App/requirements.txt
fastapi
uvicorn[standard]
torch
torchaudio
librosa
numpy
pandas
scikit-learn
python-dotenv